Uploading a boomarklet to be used on gamepress
Created a bookmarklet to allow you to quickly look up the current buffs/stats of a servant's skills. 
Added in the ability to reset / clear the currently visible window.
Added in a feature that prevents the script from embedding multiple of itself into the page. 
Future: remove from page button.