Update check_skill_bookmarklet.js
Code no longer worked due to gamepress changing the names of the classes. I have updated it to work again, and have changed some of the coloring/formatting.
Also, JQuery deprecated .size(), so I replaced all instances of it with .length. 
In the future i intend on making this a window that pops out and is draggable, but I have not done that yet.